Extract dns server to shared
