+++ packages/orchestrator/cmd/mock-sandbox/mock.go
@@ -13,10 +13,10 @@ import (
 	"go.opentelemetry.io/otel"
 	"golang.org/x/sync/errgroup"
 
-	"github.com/e2b-dev/infra/packages/orchestrator/internal/dns"
 	"github.com/e2b-dev/infra/packages/orchestrator/internal/sandbox"
 	"github.com/e2b-dev/infra/packages/orchestrator/internal/sandbox/network"
 	"github.com/e2b-dev/infra/packages/orchestrator/internal/sandbox/template"
+	"github.com/e2b-dev/infra/packages/shared/pkg/dns"
 	"github.com/e2b-dev/infra/packages/shared/pkg/grpc/orchestrator"
 	"github.com/e2b-dev/infra/packages/shared/pkg/logs"
 )
@@ -46,7 +46,7 @@ func main() {
 	go func() {
 		log.Printf("Starting DNS server")
 
-		err := dnsServer.Start("127.0.0.4:53")
+		err := dnsServer.Start("127.0.0.4", 53)
 		if err != nil {
 			log.Fatalf("Failed running DNS server: %s\n", err.Error())
 		}
